# specs/openapi.yaml
openapi: 3.1.0
info:
  title: NeatSpace API
  version: 0.1.0
  summary: Marketplace de servicios a domicilio con perfil dual (cliente/profesional).
  description: |
    Contrato REST síncrono de NeatSpace, derivado de los documentos de diseño ya
    cerrados por el panel multirol (ver `docs/00-Evaluacion-Multirol.md`).

    Reglas transversales embebidas (ver `specs/README.md`):
    - Toda escritura de dinero exige `Idempotency-Key` (400 si falta).
    - La comisión (20%) se deriva server-side; nunca viaja en el request.
    - El webhook de MercadoPago es la fuente de verdad del pago; `/pay` y `/topup` sólo inician el intento.
    - `/release` nunca libera unilateralmente (409 si no hay entrega confirmada o ventana vencida).
    - `POST /reviews` → 409 si el servicio no está pagado; el cuerpo nunca trae el 0–100.
    - Geolocalización enmascarada: dirección exacta sólo con Acuerdo aceptado o urgente asignado.
  contact:
    name: NeatSpace
  license:
    name: Propietario
servers:
  - url: https://api.neatspace.cl/v1
    description: Producción (placeholder)
  - url: http://localhost:3000/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /agreements/{id}:
    get:
      tags: [SalaDeAcuerdo]
      summary: Estado, versión vigente, partes y Trust Scores
      operationId: getAgreement
      x-source: doc 03 §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200':
          description: Acuerdo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acuerdo' }
        '404': { $ref: '#/components/responses/NotFound' }
  /agreements/{id}/versions:
    post:
      tags: [SalaDeAcuerdo]
      summary: Propone una nueva versión de términos (→ PROPUESTA)
      operationId: proposeVersion
      x-source: doc 03 §4, §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcuerdoVersionInput' }
      responses:
        '201':
          description: Nueva versión propuesta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcuerdoVersion' }
  /agreements/{id}/price-offers:
    post:
      tags: [SalaDeAcuerdo]
      summary: Contraoferta con justificación (Precio Justo)
      description: 422 si falta `justificacion` (doc 03 §11).
      operationId: makePriceOffer
      x-source: doc 03 §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceOfferInput' }
      responses:
        '201':
          description: Oferta registrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceOffer' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /agreements/{id}/price-offers/{oid}/accept:
    post:
      tags: [SalaDeAcuerdo]
      summary: Acepta una contraoferta (su monto se escribe en una nueva versión)
      operationId: acceptPriceOffer
      x-source: doc 03 §5, §11
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { name: oid, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Oferta aceptada }
        '409': { $ref: '#/components/responses/Conflict' }
  /agreements/{id}/accept:
    post:
      tags: [SalaDeAcuerdo]
      summary: Acepta la versión vigente por tu parte (requiere step-up de identidad)
      description: |
        Exige confirmación de step-up (PIN/biometría), obligatoria sobre umbral de monto
        para no repudio. Con versión desactualizada → 409. La doble aceptación gatilla la
        retención de escrow; si la retención falla, el acuerdo pasa a PAGO_FALLIDO (doc 03).
      operationId: acceptAgreement
      x-source: doc 03 §4.1, §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptInput' }
      responses:
        '200':
          description: Aceptación registrada (y escrow retenido si ambas partes aceptaron)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acuerdo' }
        '409':
          description: Versión desactualizada o pago fallido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /agreements/{id}/amendments:
    post:
      tags: [SalaDeAcuerdo]
      summary: Enmienda de alcance durante EN_EJECUCION
      description: Sólo válida en EN_EJECUCION; sobre umbral % exige evidencia/alerta NeatAI (doc 03 §11).
      operationId: amend
      x-source: doc 03 §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AmendmentInput' }
      responses:
        '201': { description: Enmienda propuesta }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /agreements/{id}/confirm:
    post:
      tags: [SalaDeAcuerdo]
      summary: El cliente confirma la entrega (→ CERRADO + libera escrow)
      description: Si vence la ventana de reclamo (24–48h), la liberación es automática (doc 03 §10).
      operationId: confirm
      x-source: doc 03 §10, §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200': { description: "Entrega confirmada, escrow liberado" }
  /agreements/{id}/cancel:
    post:
      tags: [SalaDeAcuerdo]
      summary: Cancela el acuerdo según política de estado
      operationId: cancelAgreement
      x-source: doc 03 §11
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200': { description: Cancelado; escrow resuelto según corresponda }
  /agreements/{id}/messages:
    get:
      tags: [SalaDeAcuerdo]
      summary: Mensajes de la sala
      operationId: listMessages
      x-source: doc 01 §3.3
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200':
          description: Mensajes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Mensaje' }
    post:
      tags: [SalaDeAcuerdo]
      summary: Envía un mensaje (NeatAI puede advertir sobre contacto externo)
      operationId: sendMessage
      x-source: doc 03 §7
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MensajeInput' }
      responses:
        '201': { description: Mensaje enviado }

  # ─────────────────────────── Servicios (ejecución) ───────────────────────
  /services/{id}/start:
    post:
      tags: [SalaDeAcuerdo]
      summary: Inicia la ejecución del servicio
      operationId: startService
      x-source: doc 01 §3.3
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200': { description: Servicio en ejecución }
  /services/{id}/finish:
    post:
      tags: [SalaDeAcuerdo]
      summary: Finaliza la ejecución (abre ventana de disputa)
      operationId: finishService
      x-source: doc 01 §3.1 (Flujo C)
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200': { description: Servicio finalizado }
  /services/{id}/disputes:
    post:
      tags: [SalaDeAcuerdo]
      summary: Abre una disputa (congela el escrow → EN_DISPUTA)
      operationId: openDispute
      x-source: doc 01 §3.3, doc 03 §3
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DisputaInput' }
      responses:
        '201': { description: Disputa abierta }

  # ─────────────────────────────── NeatWallet ──────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Autenticación de servicio para endpoints internos; NO alcanzables con un JWT de usuario.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: No encontrado
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflicto de estado (p. ej. versión desactualizada, primero-gana)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Entidad no procesable (validación de negocio)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: CONFLICT }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    # ── Identidad ──
...
    Acuerdo:
      type: object
      properties:
        id: { type: string }
        oportunidad_id: { type: string }
        estado:
          type: string
          enum: [ABIERTA, PROPUESTA, PAGO_FALLIDO, ACORDADO, EN_EJECUCION, ENTREGADO, EN_DISPUTA, CERRADO, CANCELADO, EXPIRADO]
        version_vigente: { $ref: '#/components/schemas/AcuerdoVersion' }
        aceptado_cliente: { type: boolean }
        aceptado_profesional: { type: boolean }
        direccion_revelada: { type: boolean, description: 'Sólo tras ACORDADO / urgente asignado' }
    AcuerdoVersionInput:
      type: object
      required: [terminos]
      properties:
        terminos: { $ref: '#/components/schemas/Terminos' }
    AcuerdoVersion:
      description: Snapshot inmutable de términos (doc 03 §4).
      type: object
      properties:
        n: { type: integer }
        terminos: { $ref: '#/components/schemas/Terminos' }
        creado_en: { type: string, format: date-time }
    Terminos:
      type: object
      properties:
        precio: { type: integer, description: 'CLP' }
        duracion: { type: string }
        materiales: { type: string, nullable: true }
        responsabilidades: { type: string, nullable: true }
        condiciones: { type: string, nullable: true }
    PriceOfferInput:
      type: object
      required: [monto, justificacion]
      properties:
        monto: { type: integer }
        justificacion: { type: string, description: 'Obligatoria (422 si falta). Estilo inDrive.' }
    PriceOffer:
      type: object
      properties:
        id: { type: string }
        monto: { type: integer }
        justificacion: { type: string }
        creado_en: { type: string, format: date-time }
    AcceptInput:
      type: object
      required: [version_n, step_up]
      properties:
        version_n: { type: integer, description: 'Versión que se acepta; 409 si ya cambió' }
        step_up:
          type: object
          description: Confirmación de identidad para no repudio (doc 03 §4.1)
          properties:
            metodo: { type: string, enum: [pin, biometria, otro] }
            token: { type: string }
    AmendmentInput:
      type: object
      required: [cambio]
      properties:
        cambio: { $ref: '#/components/schemas/Terminos' }
        evidencia_url: { type: string, nullable: true }
    DisputaInput:
      type: object
      required: [motivo]
      properties:
        motivo: { type: string }
        evidencia_url: { type: string, nullable: true }
    Mensaje:
      type: object
      properties:
        id: { type: string }
        autor_id: { type: string }
        texto: { type: string }
        leido: { type: boolean }
        alerta_neatai: { type: string, nullable: true, description: 'Advertencia de contacto externo (doc 03 §7)' }
        creado_en: { type: string, format: date-time }
    MensajeInput:
      type: object
      required: [texto]
      properties:
        texto: { type: string }

    # ── NeatWallet ──
...